refactor types into separate files
