+++ e2e/demo/application.spec.ts
@@ -4,7 +4,7 @@ import dayjs from "dayjs";
 test("create application", async ({ page }) => {
   await page.goto("/demo");
 
-  await page.getByRole("link", { name: "Create" }).click();
+  await page.getByRole("link", { name: "Create", exact: true }).click();
 
   // fill out main fields
   await page
