+++ src/app/boards/applications/create/page.tsx
@@ -1,28 +1,28 @@
 "use client";
 
-import { Controller, useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
+import Button from "@/src/components/Button";
+import ContactFields from "@/src/components/Form/ContactFields";
+import FormInput from "@/src/components/Form/FormInput";
+import FormSelectInput from "@/src/components/Form/FormSelectInput";
+import NoteFields from "@/src/components/Form/NoteFields";
+import useStorageUsage from "@/src/components/Hooks/useStorageUsage";
+import InternalLink from "@/src/components/Links/InternalLink";
+import Modal from "@/src/components/Modal";
 import {
   ApplicationStatusLabelValueType,
   applicationStatusSelectOptions,
   applicationStatuses,
   applicationStatusesArray,
   formSchema,
-} from "@/src/utils/customVariables";
-import FormInput from "@/src/components/Form/FormInput";
-import FormSelectInput from "@/src/components/Form/FormSelectInput";
+} from "@/src/types/applications";
 import { createApplication } from "@/src/utils/db";
-import ContactFields from "@/src/components/Form/ContactFields";
-import NoteFields from "@/src/components/Form/NoteFields";
-import { useEffect, useState } from "react";
-import Button from "@/src/components/Button";
-import InternalLink from "@/src/components/Links/InternalLink";
+import { zodResolver } from "@hookform/resolvers/zod";
 import dayjs from "dayjs";
-import useStorageUsage from "@/src/components/Hooks/useStorageUsage";
-import toast from "react-hot-toast";
-import Modal from "@/src/components/Modal";
 import { useSearchParams } from "next/navigation";
+import { useEffect, useState } from "react";
+import { Controller, useForm } from "react-hook-form";
+import toast from "react-hot-toast";
+import { z } from "zod";
 
 export default function Create() {
   const searchParams = useSearchParams();
