+++ src/components/Form/FormTextarea.tsx
@@ -1,8 +1,18 @@
-import {
-  FormTextareaProps,
-  defaultFocusClassName,
-} from "@/src/utils/customVariables";
+import { defaultFocusClassName } from "@/src/types/global";
+import { UseFormRegister } from "react-hook-form";
 import ErrorMessage from "./ErrorMessage";
+
+export type FormTextareaProps = {
+  id: string;
+  label: string;
+  rows?: number;
+  registerName?: string;
+  register: UseFormRegister<any>;
+  error?: string;
+  className?: string;
+  placeholder?: string;
+  isRequired?: boolean;
+};
 
 export default function FormTextarea({
   id,
