+++ src/components/Theme/ThemeProvider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import { themes } from "@/src/types/global";
 import { ThemeProvider as NextThemeProvider } from "next-themes";
-import { themes } from "../../utils/customVariables";
 
 export default function ThemeProvider({
   children,
